+++ src/compas_rhino/geometry/brep/brep.py
@@ -674,3 +674,28 @@ class RhinoBrep(Brep):
 
         """
         self._brep.Flip()
+
+    def cap_planar_holes(self, tolerance=None):
+        """Cap all planar holes in the Brep.
+
+        Parameters
+        ----------
+        tolerance : float, optional
+            The precision to use for the operation. Defaults to `TOL.absolute`.
+
+        Returns
+        -------
+        None
+
+        Raises
+        ------
+        BrepError
+            If the operation fails.
+
+        """
+        tolerance = tolerance or TOL.absolute
+        result = self._brep.CapPlanarHoles(tolerance)
+        if result:
+            self._brep = result
+        else:
+            raise BrepError("Failed to cap planar holes")